+++ Pronto-MIA/Pronto-MIA/BusinessLogic/API/Mutation.cs
@@ -33,22 +33,30 @@ namespace Pronto_MIA.BusinessLogic.API
         /// </summary>
         /// <param name="userManager">The user manager responsible for
         /// managing application users.</param>
+        /// <param name="aclManager">The manager responsible for managing
+        /// the access control lists linked to the users.</param>
         /// <param name="userName">The username of the new user.</param>
         /// <param name="password">The password of the new user.</param>
+        /// <param name="accessControlList">The access control list
+        /// that should be linked to the new user.</param>
         /// <returns>The newly created user.</returns>
         /// <exception cref="QueryException">Returns UserAlreadyExists exception
         /// if the username already exists. Alternatively returns
         /// PasswordTooWeak exception if the provided password does not meet the
         /// policy requirements.
         /// </exception>
-        [Authorize]
+        [Authorize(Policy = "CanEditUsers")]
         [UseSingleOrDefault]
         public async Task<User> CreateUser(
             [Service] IUserManager userManager,
+            [Service] IAccessControlListManager aclManager,
             string userName,
-            string password)
+            string password,
+            AccessControlList accessControlList)
         {
-            return await userManager.Create(userName, password);
+            var user = await userManager.Create(userName, password);
+            await aclManager.LinkAccessControlList(user.Id, accessControlList);
+            return user;
         }
 
         /// <summary>
@@ -61,7 +69,7 @@ namespace Pronto_MIA.BusinessLogic.API
         /// <exception cref="QueryException">Returns UserNotFound
         /// exception if the user with the given id could not be found.
         /// </exception>
-        [Authorize]
+        [Authorize(Policy = "CanEditUsers")]
         public async Task<int> RemoveUser(
             [Service] IUserManager userManager,
             int id)
@@ -75,23 +83,38 @@ namespace Pronto_MIA.BusinessLogic.API
         /// </summary>
         /// <param name="userManager">The user manager responsible for
         /// managing application users.</param>
+        /// <param name="aclManager">The manager responsible for managing
+        /// the access control lists linked to the users.</param>
         /// <param name="id">Id of the user to be updated.</param>
         /// <param name="userName">The new username of the user.</param>
         /// <param name="password">The new password for the user.</param>
         /// <returns>The updated user.</returns>
+        /// <param name="accessControlList">The new
+        /// <see cref="AccessControlList"/> that will be linked to the user.
+        /// </param>
         /// <exception cref="QueryException">Returns UserNotFound
         /// exception if the user with the given id could not be found.
         /// Alternatively returns PasswordTooWeak exception if the
         /// provided password does not meet the policy requirements.
         /// </exception>
-        [Authorize]
+        [Authorize(Policy = "CanEditUsers")]
+        [Sensitive("password")]
         public async Task<User> UpdateUser(
             [Service] IUserManager userManager,
+            [Service] IAccessControlListManager aclManager,
             int id,
             string? userName,
-            string? password)
+            string? password,
+            AccessControlList? accessControlList)
         {
-            return await userManager.Update(id, userName, password);
+            var user = await userManager.Update(id, userName, password);
+            if (accessControlList != null)
+            {
+                await aclManager.LinkAccessControlList(
+                    user.Id, accessControlList);
+            }
+
+            return user;
         }
 
         /// <summary>
@@ -108,7 +131,7 @@ namespace Pronto_MIA.BusinessLogic.API
         /// <param name="description">Short description to identify the
         /// deployment plan.</param>
         /// <returns>The newly generated deployment plan.</returns>
-        [Authorize]
+        [Authorize(Policy = "CanEditDeploymentPlans")]
         [UseSingleOrDefault]
         public async Task<IQueryable<DeploymentPlan>> CreateDeploymentPlan(
             [Service] IDeploymentPlanManager deploymentPlanManager,
@@ -144,7 +167,7 @@ namespace Pronto_MIA.BusinessLogic.API
         /// <exception cref="QueryException">Returns DeploymentPlanNotFound
         /// exception if the deployment plan with given id could not be found.
         /// </exception>
-        [Authorize]
+        [Authorize(Policy = "CanEditDeploymentPlans")]
         [UseSingleOrDefault]
         public async Task<IQueryable<DeploymentPlan>> UpdateDeploymentPlan(
             [Service] IDeploymentPlanManager deploymentPlanManager,
@@ -181,7 +204,7 @@ namespace Pronto_MIA.BusinessLogic.API
         /// <exception cref="QueryException">If the deployment plan with the
         /// given id could not be found or the firebase manager encounters
         /// a sending error.</exception>
-        [Authorize]
+        [Authorize(Policy = "CanEditDeploymentPlans")]
         [UseSingleOrDefault]
         public async Task<bool> PublishDeploymentPlan(
             [Service] IDeploymentPlanManager deploymentPlanManager,
@@ -225,7 +248,7 @@ namespace Pronto_MIA.BusinessLogic.API
         /// the deployment plan was already hidden.</returns>
         /// <exception cref="QueryException">If the deployment plan with the
         /// given id could not be found.</exception>
-        [Authorize]
+        [Authorize(Policy = "CanEditDeploymentPlans")]
         [UseSingleOrDefault]
         public async Task<bool> HideDeploymentPlan(
             [Service] IDeploymentPlanManager deploymentPlanManager,
@@ -245,41 +268,12 @@ namespace Pronto_MIA.BusinessLogic.API
         /// exception if the deployment plan with the given id could not be
         /// found.
         /// </exception>
-        [Authorize]
+        [Authorize(Policy = "CanEditDeploymentPlans")]
         public async Task<int> RemoveDeploymentPlan(
             [Service] IDeploymentPlanManager deploymentPlanManager,
             int id)
         {
             return await deploymentPlanManager.Remove(id);
-        }
-
-        /// <summary>
-        /// Sends a testing push message to the device with the given device
-        /// token.
-        /// </summary>
-        /// <param name="firebaseMessagingManager">The manager responsible for
-        /// managing firebase messaging.</param>
-        /// <param name="deviceToken">The firebase registration token of the
-        /// device the message should be sent to.</param>
-        /// <returns>True if the message could be sent false otherwise.
-        /// </returns>
-        [Authorize]
-        public async Task<bool> SendPushTo(
-            [Service] IFirebaseMessagingManager firebaseMessagingManager,
-            string deviceToken)
-        {
-            var message = new Message()
-            {
-                Notification = new Notification
-                {
-                    Title = "Test",
-                    Body = "This is a test.4",
-                },
-                Data = new Dictionary<string, string>()
-                { { "score", "850" }, { "time", "2:45" }, },
-                Token = deviceToken,
-            };
-            return await firebaseMessagingManager.SendAsync(message);
         }
 
         /// <summary>